refactor to add src/kv-bucket
