+++ src/kv-bucket/api.ts
@@ -1,0 +1,4 @@
+export interface KVBucketWithRangeQueries {
+  get (key: string, options?: { range?: { offset: number } }): Promise<Response>
+  set (key: string, value: Uint8Array): Promise<void>
+}
